extract out compute policy targets to its own function
