+++ packages/db/src/queries/index.ts
@@ -1,2 +1,3 @@
 export * from "./get-resource-parents.js";
 export * from "./create-release-job.js";
+export * from "./compute-policy-targets.js";
